Add files via upload
The first task of data cleaning is to handle simple errors in the dataset, such as missing values, formatting, and handling outliers. This task is important because it can help to improve the quality of the data and make it easier to analyze. For example, missing values can be imputed with the mean or median of the other values in the column, and outliers can be removed or winsorized.